Before vwf.FaceRecognize changes

+++ vib/cv/FaceRecognition/UIInterface.py
@@ -1,5 +1,6 @@
 import boto.sqs.connection
 from boto.sqs.connection import Message
+from boto.sqs.message import RawMessage
 import json
 import logging
 from logging import handlers
@@ -58,12 +59,15 @@ def delete_faces_for_contact( user_id, contact_id, media_asset_feature_ids ):
                 if face['face_id'] in media_asset_feature_ids:
                     delete_faces.append( face )
 
-            result = rec.delete_faces( user_id, contact_id, delete_faces )
+            if len( delete_faces ):
+                result = rec.delete_faces( user_id, contact_id, delete_faces )
 
-            if len( result['deleted'] ) != len( delete_faces ):
-                log.warning( json.dumps( { 'user_id' : user_id,
-                                           'contact_id' : contact_id, 
-                                           'message' : "Attempted to delete %s faces, successfully deleted %s faces, however %s were not found in the recognition system, and %s had errors on attempted deletion." % ( len( delete_faces ), len( result['deleted'] ), len( result['unchanged'] ), len( result['error'] ) ) } ) )
+                if len( result['deleted'] ) != len( delete_faces ):
+                    log.warning( json.dumps( { 'user_id' : user_id,
+                                               'contact_id' : contact_id, 
+                                               'message' : "Attempted to delete %s faces, successfully deleted %s faces, however %s were not found in the recognition system, and %s had errors on attempted deletion." % ( len( delete_faces ), len( result['deleted'] ), len( result['unchanged'] ), len( result['error'] ) ) } ) )
+
+        return
                 
     except Exception as e:
         log.error( json.dumps( { 'user_id' : user_id,
@@ -86,7 +90,7 @@ def move_faces( user_id, old_contact_id, new_contact_id, media_asset_feature_ids
                 if face['face_id'] in media_asset_feature_ids:
                     new_faces.append( {
                             'user_id'     : face['user_id'],
-                            'contact_id'  : face['contact_id'],
+                            'contact_id'  : new_contact_id,
                             'face_id'     : face['face_id'],
                             'face_url'    : face['face_url'],
                             'external_id' : face['external_id'],
@@ -94,18 +98,21 @@ def move_faces( user_id, old_contact_id, new_contact_id, media_asset_feature_ids
                             } )
                     delete_faces.append( face )
 
-            add_new = rec.add_faces( user_id, new_contact_id, new_faces )
+            if len( new_faces ):
+                add_new = rec.add_faces( user_id, new_contact_id, new_faces )
 
-            if len( add_new['added'] ) != len( new_faces ):
-                log.warning( json.dumps( { 'user_id' : user_id,
-                                           'message' : "Attempted to add %s faces, successfully added %s faces, however %s were already found in the recognition system, and %s had errors on attempted addition." % ( len( new_faces ), len( add_new['added'] ), len( add_new['unchanged'] ), len( add_new['error'] ) ) } ) )
+                if len( add_new['added'] ) != len( new_faces ):
+                    log.warning( json.dumps( { 'user_id' : user_id,
+                                               'message' : "Attempted to add %s faces, successfully added %s faces, however %s were already found in the recognition system, and %s had errors on attempted addition." % ( len( new_faces ), len( add_new['added'] ), len( add_new['unchanged'] ), len( add_new['error'] ) ) } ) )
 
-            delete = rec.delete_faces( user_id, old_contact_id, delete_faces )
+            if len( delete_faces ):
+                delete = rec.delete_faces( user_id, old_contact_id, delete_faces )
 
-            if len( delete['deleted'] ) != len( delete_faces ):
-                log.warning( json.dumps( { 'user_id' : user_id,
-                                           'message' : "Attempted to delete %s faces, successfully deleted %s faces, however %s were not found in the recognition system, and %s had errors on attempted deletion." % ( len( delete_faces ), len( delete['deleted'] ), len( delete['unchanged'] ), len( delete['error'] ) ) } ) )
-                            
+                if len( delete['deleted'] ) != len( delete_faces ):
+                    log.warning( json.dumps( { 'user_id' : user_id,
+                                               'message' : "Attempted to delete %s faces, successfully deleted %s faces, however %s were not found in the recognition system, and %s had errors on attempted deletion." % ( len( delete_faces ), len( delete['deleted'] ), len( delete['unchanged'] ), len( delete['error'] ) ) } ) )
+                      
+        return
     except Exception as e:
         log.error( json.dumps( { 'user_id' : user_id,
                                  'message' : "Failed to move faces: %e" % ( e ) } ) )
@@ -115,9 +122,13 @@ def __get_sqs():
     return boto.sqs.connect_to_region( config.sqs_region, aws_access_key_id = config.awsAccess, aws_secret_access_key = config.awsSecret )
 
 sqs = __get_sqs().get_queue( config.recog_queue )
+sqs.set_message_class( RawMessage )
 
 def run():
     try:
+        #import pdb
+        #pdb.set_trace()
+
         message = None
         message = sqs.read( wait_time_seconds = 20 )
 
@@ -126,6 +137,14 @@ def run():
             return True
 
         body = message.get_body()
+
+        try:
+            log.info( json.dumps( { 'message' : "Reviewing candidate message with body was: %s" % body } ) )
+        except Exception as e:
+            log.debug( json.dumps( { 'message' : "Error converting body to string, error was: %s" % e } ) )
+
+        options = json.loads( body )
+
 
         options = json.loads( body )
 
@@ -161,7 +180,7 @@ def run():
             log.error( json.dumps( { "Error, unknown action: %s" % ( action ) } ) )
 
         # DEBUG
-        #sqs.delete_message( message )
+        sqs.delete_message( message )
         return True
 
     except Exception as e:
